сделал дело, гуляй смело

+++ src/main/java/ru/kstu/aec/models/Answer.java
@@ -15,9 +15,9 @@ public class Answer {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
-    private int question_id;
+    private Long question_id;
 
     private String answer_text;
